requirements.txt
flask
python-dotenv
google-generativeai
sentence-transformers
chromadb
PyMuPDF
requests